+++ fe/src/constants/room.ts
@@ -1,0 +1,3 @@
+export const ROOM_RULES = Object.freeze({
+  maxPage: 9,
+});
